refactor(common): 폴더 명 변경 (dto -> response)
